% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{insert_mucic_idx}
\alias{insert_mucic_idx}
\title{Index para parear dados}
\usage{
insert_mucic_idx(x, vec = seq(645))
}
\arguments{
\item{x}{e um vetor qualquer, precisamos do tamanho dele}

\item{vec}{e o vetor de indexacao (1:645 municipios) que sera replicado
\code{x} vezes}
}
\value{
uma \code{list} com dois vetores igual a \code{vec}.
}
\description{
Index para parear dados
}
