% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{seade}
\alias{seade}
\title{Fundacao Seade}
\format{a `tibble` with 40 rows and 2 colums
\describe{
  \item{ano}{years from 1980 to 2017}
  \item{data}{a vector os \code{lists} with nested data}
}}
\usage{
seade
}
\description{
Socio economic statistcs of Sao Paulo estate, from SEADE foundation.
}
\keyword{datasets}
